Give a URL one identity, and stop letting a malformed one crash the audit
There were four notions of URL identity in the package and they disagreed.
`same_site` folded `www.` while `normalize_url` did not, so a link written in the
`www.` form passed the same-site filter, got a different dedup identity, and was
fetched twice: the corpus then held one document under two names and a healthy
site failed ADS-CONTENT-02 with "2 of 4 analyzed pages (50%)". `completeness`
compared netloc verbatim in a third place, so one broken navigation URL linked
under both spellings counted twice, and three counted links is the line between
WARNING and FAIL. There is now one `_netloc_identity`; `site_host`,
`normalize_url` and `completeness._canonical` all defer to it.

Five ways it was wrong about what a site is:

  * the port was compared against the default of that URL's own scheme, so
    `http://host:443` and `https://host` — what a TLS-terminating proxy emits
    when it reports SERVER_PORT=443 without an HTTPS flag — were two sites, and a
    declared About page written that way was never requested while the report
    said "no candidate answered (7 tried)" without it;
  * an IPv6 host came back unbracketed, so `[::1]:9411` and `[::1:9411]` were one
    identity: a third party's address was crawled as the audited site's own page,
    under a robots.txt never read for it;
  * a trailing-dot FQDN was another site, and every link written that way went to
    `off_site` — the "crawled 1 URL and reported no problem" failure;
  * an internationalised host and its punycode were two sites, though requests
    IDNA-encodes at fetch time and both reach the same server;
  * an unparseable URL rendered as `http:///<path>`, giving every unparseable URL
    with that path one identity, so two distinct off-site links collapsed into one
    and the second was never scanned or reported.

`urlsplit` raises ValueError on a malformed bracketed host, and ValueError is not
a RequestException, so it walked out of `fetch` and out of every URL helper: four
of the five CLIs died on it, two of them needing nothing more than one bad href
on the audited page. `split_url`, `join_url` and `defrag_url` never raise.
`urldefrag` parses too, which is why it took two passes to find: it was still raw
in `normalize_url` and in `_resolve_links`, where a `<base href>` that fails to
join hands back the unparsed href. And `fetch` no longer catches ValueError
wholesale — it validates the URL first, so a caller passing `timeout=0` gets
their own exception instead of a verdict about a site that was up all along.

Three more the identity work uncovered:

  * the crawler sent `user:pass@` credentials scraped from the page it was
    auditing, so the 401 never happened and "readable publicly, without
    authentication" passed on a page no anonymous visitor can open;
  * `head` sat in the crawler's non-text tag set, and HTML5 lets a document omit
    `</head>`, which HTMLParser does not synthesise — so the counter never came
    back down and a 300-word page reported 0 visible words. The other two parsers
    in the package read that document correctly;
  * `extra_seeds` were queued with no same-site check and matched against the
    audited site's robots.txt, the one place this tool would fetch a third
    party's URL under a file never read for it.

Also here, because they live in these two modules: `as_base` appended a slash
unconditionally, so auditing `…/index.html` asked for `…/index.html/` and the
whole run reported "Home page could not be read: HTTP 404" about a site that was
up; a crawl truncated by `--max-pages` printed `[PASS] pages reachable` while
three 500s sat just past the ceiling; and navigation links past `--nav-limit`
were not looked at, which is MISSING and not a pass.

+++ adsense_checks/duplicates.py
@@ -534,7 +534,7 @@ def fetch_page_texts(
     urls: Iterable[str],
     *,
     extract: Callable[[str], str] | None = None,
-    timeout: int = DEFAULT_TIMEOUT,
+    timeout: float = DEFAULT_TIMEOUT,
     session: requests.Session | None = None,
 ) -> Corpus:
     """Fetch each URL and extract its text, recording why any page is missing.
@@ -634,7 +634,7 @@ def check_urls(
     threshold: float = DEFAULT_SIMILARITY_THRESHOLD,
     page_ratio_threshold: float = DEFAULT_DUPLICATE_PAGE_RATIO,
     extract: Callable[[str], str] | None = None,
-    timeout: int = DEFAULT_TIMEOUT,
+    timeout: float = DEFAULT_TIMEOUT,
     session: requests.Session | None = None,
     shingle_size: int = SHINGLE_SIZE,
 ) -> DuplicationResult:
